requirements.txt
Flask==2.3.3
PyMuPDF==1.23.7
langdetect==1.0.9
pytesseract==0.3.10
pillow==10.2.0
